:goal_net: Added error handlers

+++ src/main/java/com/que/quevent/services/AttendeeService.java
@@ -1,5 +1,7 @@
 package com.que.quevent.services;
 
+import com.que.quevent.exeptions.AttendeeAlreadyExistsException;
+import com.que.quevent.exeptions.AttendeeNotExistsException;
 import com.que.quevent.model.Attendee;
 import com.que.quevent.repositories.AttendeeRepository;
 import org.springframework.stereotype.Service;
@@ -14,6 +16,9 @@ public class AttendeeService {
     }
 
     public void addNewAttendee(Attendee attendee) {
+        if (attendeeRepository.existsById(attendee.getId())) {
+            throw new AttendeeAlreadyExistsException("Такой участник уже существует");
+        }
         attendeeRepository.save(attendee);
     }
 
@@ -22,10 +27,17 @@ public class AttendeeService {
     }
 
     public void deleteAttendee(Attendee attendee) {
-        attendeeRepository.delete(attendee);
+        if (attendeeRepository.existsById(attendee.getId())) {
+            attendeeRepository.delete(attendee);
+        }
+        throw new AttendeeNotExistsException("Такого участника не существует");
     }
 
     public void deleteAttendeeById(long id) {
-        attendeeRepository.deleteById(id);
+        if (attendeeRepository.existsById(id)) {
+            attendeeRepository.deleteById(id);
+        }
+        throw new AttendeeNotExistsException("Такого участника не существует");
+
     }
 }
